move BellPairStore into its directory
