+++ quisp/modules/QRSA/RuleEngine/RuleEngine.h
@@ -13,7 +13,7 @@
 
 #include <messages/classical_messages.h>
 #include "../../PhysicalConnection/BSA/HoMController.h"
-#include "BellPairStore.h"
+#include "BellPairStore/BellPairStore.h"
 #include "IRuleEngine.h"
 #include "modules/QNIC/StationaryQubit/IStationaryQubit.h"
 #include "modules/QRSA/HardwareMonitor/HardwareMonitor.h"
